Minor refactoring for clarity: combined benchmark classes into single module, removed unused files, updated README, renamed dirs to emphasize python benchmarks.
