--- benchmarks/end_to_end/README.md
@@ -1,55 +1,0 @@
-# cuGraph SNMG benchmarks
-
-## Overview
-
-This directory contains source and configuration files for benchmarking
-`cuGraph`.  The sources are currently intended to benchmark `cuGraph` via the
-python API, but this is not a requirement, and future updates may include
-benchmarks written in C++ or other languages
-
-The benchmarks here use datasets generated by the RMAT but also support csv
-files as input
-
-## Prerequisites
-### Python
-* cugraph built and installed (or `cugraph` sources and built C++ extensions
-  available on `PYTHONPATH`)
-
-## Usage (Python)
-### Python
-* Run `python main.py --help` for a list of all available benchmark options
-
-* Expose the desired GPUs to the benchmark run via CUDA_VISIBLE_DEVICES
-
-* Run `bash  ./run_all_nightly_benches.sh` for automated benchmarks with multiple scales,
-gpu configuration and edgefactor
- 
-## Examples
-### Python
-_**NOTE: Some algos require the graph to be symmetrize (Louvain, WCC) or unweighted.**_
-* Run all the benchmarks with a generated datasets of scale=23
-```
-(rapids) user@machine:/cugraph/benchmarks/end_to_end> python nightly/main.py --scale=23
-```
-
-* Run all the benchmarks with a generated unweighted dataset of scale=23
-```
-(rapids) user@machine:/cugraph/benchmarks/end_to_end> python nightly/main.py --scale=23 --unweighted
-```
-
-* Symmetrize the generated dataset of scale=23 and run all the benchmarks
-```
-(rapids) user@machine:/cugraph/benchmarks/end_to_end> python nightly/main.py --scale=23 --symmetric-graph
-```
-
-* Create a graph from a csv file an run all the benchmarks
-```
-(rapids) user@machine:/cugraph/benchmarks/end_to_end> python nightly/main.py --csv='karate.csv'
-```
-
-* Run unweighted algos(Pagerank, Katz) with multiple scales(23~30) and gpus configuration(1, 2, 4 and 8 GPUs)
-```
-(rapids) user@machine:/cugraph/benchmarks/end_to_end> bash ./run_all_nightly_benches.sh
-```
-
-
