+++ benchmarks/python_e2e/README.md
@@ -1,0 +1,73 @@
+# cuGraph benchmarks
+
+## Overview
+
+This directory contains source and configuration files for benchmarking
+`cuGraph`.  The sources are currently intended to benchmark `cuGraph` via the
+python API, but this is not a requirement, and future updates may include
+benchmarks written in C++ or other languages
+
+The benchmarks here use datasets generated by the RMAT graph generator but also
+support csv files as input
+
+## Prerequisites
+### Python
+* cugraph built and installed (or `cugraph` sources and built C++ extensions
+  available on `PYTHONPATH`)
+
+## Usage
+* Run `python main.py --help` for a list of all available benchmark options
+
+* Expose the desired GPUs to the benchmark run via CUDA_VISIBLE_DEVICES
+
+* Run `bash ./run_all_nightly_benches.sh` for automated benchmarks with multiple
+  scales, gpu configuration and edgefactor
+
+* Run `python main.py` to run individual algo benchmarks with specific
+  options. For example, to benchmark a 2-GPU run of BFS and SSSP with a
+  generated graph of size scale 23:
+```
+(rapids) user@machine:/cugraph/benchmarks/python_e2e> export CUDA_VISIBLE_DEVICES=0,1
+
+(rapids) user@machine:/cugraph/benchmarks/python_e2e> python main.py --scale=23 --algo=bfs --algo=sssp
+calling setup...distributed.preloading - INFO - Import preload module: dask_cuda.initialize
+distributed.preloading - INFO - Import preload module: dask_cuda.initialize
+done.
+running generate_edgelist (RMAT)...done.
+running from_dask_cudf_edgelist...done.
+running compute_renumber_edge_list...done.
+running compute_renumber_edge_list...done.
+running bfs (warmup)...done.
+running bfs...done.
+running sssp (warmup)...done.
+running sssp...done.
+from_dask_cudf_edgelist()     0.0133009
+------------------------------------------------------------
+bfs(start:73496080)           0.569328
+------------------------------------------------------------
+sssp(start:73496080)          1.48114
+
+calling teardown...done.
+```
+
+## Other Examples:
+_**NOTE: Some algos require the graph to be symmetrized (Louvain, WCC) or unweighted.**_
+* Run all the benchmarks with a generated datasets of scale=23
+```
+(rapids) user@machine:/cugraph/benchmarks/python_e2e> python main.py --scale=23
+```
+
+* Run all the benchmarks with a generated unweighted dataset of scale=23
+```
+(rapids) user@machine:/cugraph/benchmarks/python_e2e> python main.py --scale=23 --unweighted
+```
+
+* Symmetrize the generated dataset of scale=23 and run all the benchmarks
+```
+(rapids) user@machine:/cugraph/benchmarks/python_e2e> python main.py --scale=23 --symmetric-graph
+```
+
+* Create a graph from a csv file an run all the benchmarks
+```
+(rapids) user@machine:/cugraph/benchmarks/python_e2e> python main.py --csv='karate.csv'
+```
