+++ benchmarks/python_e2e/benchmark.py
@@ -11,13 +11,55 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
-from benchmark import benchmark
-
 import sys
-def log(s, end="\n"):
-    print(s, end=end)
-    sys.stdout.flush()
+import time
+from functools import wraps
+
+
+class BenchmarkedResult:
+    """
+    Class to hold results (the return value of the callable being benchmarked
+    and meta-data about the benchmarked function) of a benchmarked function run.
+    """
+    def __init__(self, name, retval, runtime, params=None):
+        self.name = name
+        self.retval = retval
+        self.runtime = runtime
+        self.params = params or {}
+        self.validator_result = True
+
+
+def benchmark(func):
+    """
+    Returns a callable/closure that wraps func with code to time the func call
+    and return a BenchmarkedResult. The resulting callable takes the same
+    args/kwargs as func.
+
+    The BenchmarkedResult will have its params value assigned from the kwargs
+    dictionary, but the func positional args are not captured. If a user needs
+    the params captured for reporting purposes, they must use kwargs.  This is
+    useful since positional args can be used for args that would not be
+    meaningful in a benchmark result as a param to the benchmark.
+
+    This can be used as a function decorator or a standalone function to wrap
+    functions to benchmark.
+    """
+    benchmark_name = getattr(func, "benchmark_name", func.__name__)
+    @wraps(func)
+    def benchmark_wrapper(*func_args, **func_kwargs):
+        t1 = time.perf_counter()
+        retval = func(*func_args, **func_kwargs)
+        t2 = time.perf_counter()
+        return BenchmarkedResult(name=benchmark_name,
+                                 retval=retval,
+                                 runtime=(t2-t1),
+                                 params=func_kwargs,
+                                )
+
+    # Assign the name to the returned callable as well for use in debug prints,
+    # etc.
+    benchmark_wrapper.name = benchmark_name
+    return benchmark_wrapper
 
 
 class BenchmarkRun:
@@ -63,16 +105,22 @@ class BenchmarkRun:
         self.results = []
 
 
+    @staticmethod
+    def __log(s, end="\n"):
+        print(s, end=end)
+        sys.stdout.flush()
+
+
     def run(self):
         """
         Run and time the graph construction step, then run and time each algo.
         """
         self.results = []
 
-        log(f"running {self.construct_graph.name}...", end="")
+        self.__log(f"running {self.construct_graph.name}...", end="")
         result = self.construct_graph(self.input_dataframe,
                                       *self.construct_graph_func_args)
-        log("done.")
+        self.__log("done.")
         G = result.retval
         self.results.append(result)
 
@@ -94,18 +142,18 @@ class BenchmarkRun:
                 if hasattr(G, "compute_renumber_edge_list"):
                     G.compute_renumber_edge_list(transposed=True)
             else: #set transpose=False when renumbering
-                log("running compute_renumber_edge_list...", end="")
+                self.__log("running compute_renumber_edge_list...", end="")
                 if hasattr(G, "compute_renumber_edge_list"):
                     G.compute_renumber_edge_list(transposed=False)
-                log("done.")
+                self.__log("done.")
         # FIXME: need to handle individual algo args
         for ((algo, params), validator) in zip(self.algos, self.validators):
-            log(f"running {algo.name} (warmup)...", end="")
+            self.__log(f"running {algo.name} (warmup)...", end="")
             algo(G, **params)
-            log("done.")
-            log(f"running {algo.name}...", end="")
+            self.__log("done.")
+            self.__log(f"running {algo.name}...", end="")
             result = algo(G, **params)
-            log("done.")
+            self.__log("done.")
 
             if validator:
                 result.validator_result = validator(result.retval, G)
